requirements.txt
# © 2025 Sumedh Sohan - JARVIS AI Assistant
# All rights reserved. Unauthorized use is strictly prohibited.

flask==2.3.3
flask-cors==4.0.0
mistralai==0.4.2
psutil==5.9.6
requests==2.31.0
python-dotenv==1.0.0
pyautogui==0.9.54
pynput==1.7.6
pillow==10.0.1
opencv-python==4.8.1.78
SpeechRecognition==3.10.0
pyttsx3==2.90
schedule==1.2.0
plyer==2.1.0